refactor(gateway): modularize main.rs and prune over-exposure
The bin-only gateway crate had a ~3,950-line main.rs holding the crate-root
glue, the entire server engine, and the CLI. Split it into focused modules and
tighten visibility now that nothing outside the crate consumes it:

- main.rs (352): crate root — module declarations, the Shared* state aliases
  (referenced by the engine and submodules), and the `#[tokio::main]` entry.
- server.rs (3,271): the server engine — run_gateway, the connection/dispatch
  loop, model orchestration.
- cli.rs (263): clap argument definitions and the `pair` subcommand handler.

Over-exposure cleanup (the crate exports no public API):
- Demoted every `pub mod` to `mod` and dropped the now-pointless `pub use`
  re-exports.
- Removed four modules that were never wired into the gateway and never used
  anywhere in the workspace (dead since creation, previously masked by being
  `pub` in the old core lib): csrf, health, thinking_clock, webhooks (~1,130
  lines).
- Removed `run_stdio_gateway` — dead code; the `--ssh-stdio` path is handled
  inline in `run_gateway` via `handle_transport_connection`.
- Annotated genuinely partial-but-unwired features with `#[allow(dead_code)]`
  and a note explaining the pending wiring: mcp_handler / canvas_handler tool
  execution (detection halves are wired; execution needs the manager/host
  threaded through — see the TODO markers in messenger_handler), the
  concurrent producer side (ChannelSink / ActiveTasks), a few GatewayError
  variants, and unread SshConfig fields.

cargo build --workspace, --no-default-features, and
-p rustyclaw-gateway --features all-messengers: clean.
cargo clippy --workspace --all-targets -- -D warnings: clean.
cargo fmt --all -- --check: clean.
